test: refactor: move `CreateDescriptor` helper to wallet test util module
Can be reviewed via `--color-moved=dimmed-zebra`.
